Remove check for childRestObjs size
The check for childRestObjs size prevents a user from un-assigning all objects that are assigned to a given object. An example where this occurs is is with and PolicyGroup the VPorts that are assigned to it. Because of this check in RestObject, I cannot un-assign all VPorts from a PolicyGroup, thus I can't delete a PolicyGroup. Another example is PatNatPools assigned to a Vlan.
This code change has not been tested, as I am currently having issues cloning this repo.